feat: add event search component for selecting and inviting collaborators
- Implemented EventSearch component with SelectMode and InviteMode for searching and inviting collaborators.
- Integrated hooks for campus and IG searches, and collaboration targets.
- Updated constants for event types and collaborator types.
- Refactored event detail view to improve edit history display.
- Removed deprecated badge constants and consolidated event-related constants into a single file.
- Enhanced user search input by simplifying normalization logic.
- Improved event hooks for better data handling and query management.
