+++ src/features/events/components/index.ts
@@ -1,10 +1,9 @@
 export * from "./co-owners-panel";
-export * from "./collaborator-search-input";
 export * from "./collaborators-panel";
-export * from "./entity-search-select";
 export * from "./event-card";
 export * from "./event-detail-view";
 export { default as EventModal } from "./event-modal";
+export * from "./event-search";
 export * from "./event-status-badge";
 export * from "./event-type-badge";
 export * from "./events-filters";
