+++ src/features/events/constants/index.ts
@@ -1,2 +1,1 @@
-export * from "./badge.constants";
-export * from "./events.constants";
+export * from "./constants";
